--- src/components/app-header/app-header.module.css
@@ -1,84 +1,0 @@
-
-
-  .app_header {
-    display: flex;
-    flex-flow: row nowrap;
-   /*justify-content: space-around;*/
-   justify-content: space-evenly;
-   flex-direction: row;
-   /*justify-content: center;*/
-  
- 
-    height: 88px;
-    left: 2%;
-    right: 2%;
-    top: 40px;
-
-    /* interface/modal-bg */
-
-    background: #1C1C21;
-    /* interface/accent */
-
-    box-shadow: 0px 24px 32px rgba(0, 0, 0, 0.04), 0px 16px 24px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.04), 0px 0px 1px rgba(0, 0, 0, 0.04);
-}
-
-.navigation_panel{
-
-
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 16px 20px;
-  gap: 8px;
-
-  justify-content: left;
-
-  flex-basis: 30%;
-
-}
-
-
-.navigation_link{
-
-
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 16px 20px;
-    gap: 8px;
-
-
-
-}
-
-
-.navigation_logo{
-
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-basis: 30%;
-
-
-}
-
-
-.navigation_account{
-
-
-  display: flex;
-  justify-content: right;
-
-  flex-basis: 30%;
-
-
-}
-
-  
-.flex_center {
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}
-
