--- src/components/app/app.module.css
@@ -1,20 +1,0 @@
-.app {
-  text-align: center;
-
-}
-
-
-.app_container {
-
-  min-height: 100vh;
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: stretch;
-  justify-content: center;
-  
-
-  gap: 40px;
-}
-
-
-
