--- src/components/burger-ingredient-card/burger-ingredient-card.module.css
@@ -1,23 +1,0 @@
-.burger_ingredient_card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    position: relative;
-
-    width: 260px;
-    height: 200px;
-
-    margin: 4px 4px 0px 0px;
-
-   /* border: 1px dashed blue;
-    box-sizing: border-box;*/
-  }
-
-  .flex_center_row{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    gap: 10px;
-}
