--- src/components/burger-ingredient-category/burger-ingredient-category.module.css
@@ -1,32 +1,0 @@
-
-.burger_ingredient_category {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    margin: 0px 0px 20px 0px;
-
-  /*  border: 1px dashed red;
-    box-sizing: border-box;*/
-  }
-
-  .burger_ingredient_category_caption {
-    width: 100%;
-    text-align: left;
-
-
-  /*  border-style: dashed;  
-    border-color: green;   */
-  }
-
-  .burger_ingredient_list{
-
-    display: flex;
-    align-items: flex-start;
-    flex-wrap: wrap;
-    gap: 20px;
-    margin-left: 16px; 
-
-}
-
-
